Added core for Kermit
