+++ .idea/artifacts/crashkios_core_jvm_0_4_0.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="crashkios-core-jvm-0.4.0">
+    <output-path>$PROJECT_DIR$/crashkios-core/build/libs</output-path>
+    <root id="archive" name="crashkios-core-jvm-0.4.0.jar">
+      <element id="module-output" name="CrashKiOS.crashkios-core.jvmMain" />
+    </root>
+  </artifact>
+</component>
