+++ gradle.properties
@@ -3,7 +3,7 @@ kotlin.code.style=official
 
 #kotlin.native.home=/Users/kgalligan/temp/kotlin-native-master-hold/dist
 GROUP=co.touchlab
-VERSION_NAME=0.3.2
+VERSION_NAME=0.4.0
 
 POM_URL=https://github.com/touchlab/CrashKios
 POM_DESCRIPTION=Kotlin Native iOS Crash Report Library
@@ -21,3 +21,7 @@ POM_DEVELOPER_ID=kpgalligan
 POM_DEVELOPER_NAME=Kevin Galligan
 POM_DEVELOPER_ORG=Kevin Galligan
 POM_DEVELOPER_URL=https://touchlab.co/
+
+kotlin.mpp.enableGranularSourceSetsMetadata=true
+kotlin.native.enableDependencyPropagation=false
+kotlin.mpp.commonizerLogLevel=info
